move evaluate_function_over_domain into sexe-expression
